add messaging areas
User Messaging Area (UMA) is only available on the CT80. I believe that Price Messaging Area (PMA) is available for most all models so I put it in the CommonThermostat class but I was only able to test against the CT80. The API allows 4 lines for the PMA but I could only make 2 work, may be different per model?